read
a) Hard to tell with all that junk code in there but it looks a bit on the light side now.
b) Yes, I see some tests, just get rid of all that junk in the file.
c) main.ml was in the wrong spot so build file would not work.
Why do you have old assignment code in there? Makes things very hard to read.
Is there a reason why valid_move is in lib.ml? Seems like it is more a board property.
-5
d) I would not alias move to int. If you are wanting to do that make a variant something like `type move = Move of int` so it is explicit.
getPos would be better taking a board position (pair of ints) as argument.
Make player not be an int, they are just tokens so can be variants like Player1, Player2.
write a separate function to check for "xPos < 0 || xPos > 6 || yPos < 0 || yPos > 5" kind of stuff, is_valid_position kind of thing.
isGameOver should be is_game_over, see Course style guide.
-7